fix(review): 修复时间轴边界与末窗口采样并兼容 Python 3.10
验证源时间零起点与末帧、清单真实帧数、评估骨架及空集合；数据清单哈希证明划分一致性。GIF 按源帧间隔播放。分块计算权重哈希，包含最后合法训练窗口。23 项 Python 回归和 UE 模块重编译通过。

+++ Unreal/AILocomotionSystem/AILocomotionDataset/Source/AILocomotionDatasetEditor/Private/AILocomotionDatasetEditorModule.cpp
@@ -56,7 +56,7 @@ namespace
 		}
 		const double Duration = Animation->GetPlayLength();
 		const FFrameRate SourceRate = DataModel->GetFrameRate();
-		const double FrameCount = DataModel->GetNumberOfKeys();
+		const int32 FrameCount = DataModel->GetNumberOfKeys();
 		if (!FMath::IsFinite(Duration) || FrameCount < 2 || FrameCount > Settings->MaxFramesPerClip || FrameCount > 18000)
 		{
 			Error = LOCTEXT("InvalidDuration", "动画长度超出采样范围，请检查单片段最大帧数。").ToString();
@@ -162,7 +162,7 @@ namespace
 		Options.bIncorporateRootMotionIntoPose = true;
 		Options.bEvaluateCurves = false;
 		TArray<TSharedPtr<FJsonValue>> Frames;
-		for (int32 FrameIndex = 0; FrameIndex < static_cast<int32>(FrameCount); ++FrameIndex)
+		for (int32 FrameIndex = 0; FrameIndex < FrameCount; ++FrameIndex)
 		{
 			Progress.EnterProgressFrame(1.0f / static_cast<float>(FrameCount));
 			if (Progress.ShouldCancel())
